+++ docs/changelog.md
@@ -30,7 +30,12 @@ See [upgrade notes][1] for helpful information when upgrading from previous vers
 
 ## Unreleased
 
+What's changed since v1.47.0:
+
 - New rules:
+  - Azure Container Registry:
+    - Check that audit diagnostic logs are enabled for Container Registry by @copilot.
+      [#3536](https://github.com/Azure/PSRule.Rules.Azure/issues/3536)
   - Container Apps:
     - Check that liveness and readiness health probes use HTTP checks for HTTP-based ingress.
       [#3714](https://github.com/Azure/PSRule.Rules.Azure/issues/3714)